+++ framework/src/main/java/org/checkerframework/framework/util/JavaParserUtil.java
@@ -12,6 +12,7 @@ import com.github.javaparser.ast.body.ClassOrInterfaceDeclaration;
 import com.github.javaparser.ast.body.EnumDeclaration;
 import com.github.javaparser.ast.body.TypeDeclaration;
 import com.github.javaparser.ast.expr.AnnotationExpr;
+import com.github.javaparser.ast.expr.ArrayInitializerExpr;
 import com.github.javaparser.ast.expr.BinaryExpr;
 import com.github.javaparser.ast.expr.Expression;
 import com.github.javaparser.ast.expr.StringLiteralExpr;
@@ -266,6 +267,11 @@ public class JavaParserUtil {
         }
       }
     }
+
+    @Override
+    public void visit(ArrayInitializerExpr node, Void p) {
+      // Do not remove annotations that are array elements.
+    }
   }
 
   /**
@@ -358,8 +364,8 @@ public class JavaParserUtil {
         case "RELEASE_17":
           currentSourceVersion = ParserConfiguration.LanguageLevel.JAVA_17;
           break;
-          // As of 2022-09-27, JavaParser's ParserConfiguration.LanguageLevel does not yet have a
-          // constant for JDK 18.
+          // As of 2022-11-28, JavaParser's ParserConfiguration.LanguageLevel does not yet have a
+          // constant for JDK 18 or JDK 19.
           // case "RELEASE_18":
           //   currentSourceVersion = ParserConfiguration.LanguageLevel.JAVA_18;
           //   break;